refactor: rename 'data_type' module with 'argument'
