+++ arbintoolbox/test/arbincti/feedback/test_schedule_operation.py
@@ -8,7 +8,7 @@ from System.Collections.Generic import ( # type: ignore
 )
 import ArbinCTI.Core as ArbinCTI # type: ignore
 
-from arbintoolbox.src.arbincti.data_type.cti_data_type import (
+from arbintoolbox.src.arbincti.argument.argument import (
     TE_DATA_TYPE,
     TimeSensitiveSetMV
 )
